[libs-core] refactor: BlockFile 클래스 - block 패키지 리팩토링 영향 반영 (#29)
