+++ libs/core/src/main/java/kr/joberchip/core/storage/ProfileImageFile.java
@@ -1,7 +1,8 @@
 package kr.joberchip.core.storage;
 
 import javax.persistence.*;
-import kr.joberchip.core.space.page.SpacePageProfile;
+
+import kr.joberchip.core.share.page.SharePage;
 
 @Entity
 @Table(name = "profile_image_file_tb")
@@ -11,7 +12,7 @@ public class ProfileImageFile {
   @Column(name = "profile_image_file_id")
   private Long profileImageFileId;
 
-  @OneToOne private SpacePageProfile pageProfile;
+  @OneToOne private SharePage sharePage;
 
   @OneToOne private AttachedFile attachedFile;
 }
